+++ Common/src/main/java/fuzs/puzzleslib/api/core/v1/CommonAbstractions.java
@@ -13,7 +13,7 @@ public interface CommonAbstractions {
     /**
      * instance of the common abstractions SPI
      */
-    CommonAbstractions INSTANCE = ServiceProviderHelper.loadServiceProvider(CommonAbstractions.class);
+    CommonAbstractions INSTANCE = ServiceProviderHelper.load(CommonAbstractions.class);
 
     /**
      * opens a menu on both client and server