+++ Common/src/main/java/fuzs/puzzleslib/impl/client/core/ClientFactories.java
@@ -8,7 +8,7 @@ import fuzs.puzzleslib.api.core.v1.ServiceProviderHelper;
 import java.util.function.Supplier;
 
 public interface ClientFactories {
-    ClientFactories INSTANCE = ServiceProviderHelper.loadServiceProvider(ClientFactories.class);
+    ClientFactories INSTANCE = ServiceProviderHelper.load(ClientFactories.class);
 
     void constructClientMod(String modId, Supplier<ClientModConstructor> modConstructor, ContentRegistrationFlags... contentRegistrations);