+++ Common/src/main/java/fuzs/puzzleslib/impl/config/AnnotatedConfigBuilder.java
@@ -5,39 +5,23 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.ObjectArrays;
-import fuzs.puzzleslib.api.config.v3.ConfigCore;
 import fuzs.puzzleslib.api.config.v3.Config;
+import fuzs.puzzleslib.api.config.v3.ConfigCore;
 import net.minecraftforge.common.ForgeConfigSpec;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import sun.misc.Unsafe;
 
-import java.lang.reflect.Constructor;
+import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.*;
 
 /**
- * build config values from given class, values are marked via {@link Config} annotation
- * large based upon <a href="https://github.com/VazkiiMods/Quark">Quark's</a> <code>vazkii.quark.base.module.config.ConfigObjectSerializer</code> class
+ * Build config values from given class, values are marked via {@link Config} annotation.
+ * <p>Large based upon <a href="https://github.com/VazkiiMods/Quark/blob/master/src/main/java/vazkii/quark/base/module/config/ConfigObjectSerializer.java">ConfigObjectSerializer.java</a> from the Quark mod.
  */
 public class AnnotatedConfigBuilder {
-    /**
-     * Unsafe instance for setting final fields in configs, so they may remain effectively read-only.
-     */
-    private static final Unsafe UNSAFE;
-
-    static {
-        try {
-            Constructor<?> constructor = Unsafe.class.getDeclaredConstructors()[0];
-            constructor.setAccessible(true);
-            UNSAFE = (Unsafe) constructor.newInstance();
-        }
-        catch (Throwable e) {
-            throw new RuntimeException(e);
-        }
-    }
 
     /**
      * @param builder forge builder for creating forge config values, setting comments, etc.
@@ -82,8 +66,6 @@ public class AnnotatedConfigBuilder {
                 field.setAccessible(true);
                 final boolean isStatic = Modifier.isStatic(field.getModifiers());
                 if (!isStatic) Objects.requireNonNull(instance, "Null instance for non-static field");
-                // final fields are enforced since config values are read-only
-                if (!Modifier.isFinal(field.getModifiers())) throw new RuntimeException("Field must be final");
                 buildConfig(builder, context, isStatic ? null : instance, field, field.getDeclaredAnnotation(Config.class));
             }
             if (!path.isEmpty()) builder.pop(path.size());
@@ -101,7 +83,7 @@ public class AnnotatedConfigBuilder {
      */
     private static Map<List<String>, Collection<Field>> setupFields(Class<?> target) {
         Multimap<List<String>, Field> pathToField = HashMultimap.create();
-        for (Field field : collectFieldsRecursive(target)) {
+        for (Field field : getAllFieldsRecursive(target)) {
             Config annotation = field.getDeclaredAnnotation(Config.class);
             if (annotation != null) {
                 pathToField.put(Lists.newArrayList(annotation.category()), field);
@@ -116,7 +98,7 @@ public class AnnotatedConfigBuilder {
      * @param clazz class to get fields for
      * @return all fields from class and all superclasses
      */
-    private static List<Field> collectFieldsRecursive(Class<?> clazz) {
+    private static List<Field> getAllFieldsRecursive(Class<?> clazz) {
         List<Field> list = new LinkedList<>();
         while (clazz != Object.class) {
             list.addAll(Arrays.asList(clazz.getDeclaredFields()));
@@ -159,6 +141,10 @@ public class AnnotatedConfigBuilder {
             builder.pop();
             return;
         }
+
+        // final fields are permitted until here, since values must be able to change
+        // previously only new config categories are handled, those instances never change
+        if (Modifier.isFinal(field.getModifiers())) throw new RuntimeException("Field may not be final");
 
         // does this require a world restart? just an indicator, doesn't actually do something
         if (annotation.worldRestart()) builder.worldRestart();
@@ -246,9 +232,12 @@ public class AnnotatedConfigBuilder {
      * @param instance object instance, null when static
      */
     private static void addCallback(ConfigDataHolderImpl<?> context, ForgeConfigSpec.ConfigValue<?> configValue, Field field, @Nullable Object instance) {
-        context.accept(configValue, v -> {
-            long fieldOffset = UNSAFE.objectFieldOffset(field);
-            UNSAFE.putObject(instance, fieldOffset, configValue.get());
+        context.accept(configValue, value -> {
+            try {
+                MethodHandles.lookup().unreflectSetter(field).invoke(instance, configValue.get());
+            } catch(Throwable e) {
+                throw new RuntimeException(e);
+            }
         });
     }
 }